Fix input bug A newline input doesn't result in an error anymore.

+++ util/delnote.sh
@@ -11,16 +11,24 @@ function delnote() {
 	else
 		delete=$1
 	fi
+
+	if [ -z $delete ]; then
+		echo "Invalid input."
+		return 0
+
+	else
 		if [[ $1 == "-all" ]]; then
-			cd $DEFAULT_PATH
-			rm -f *
-			clear
-			cd $DIR
-			str="All files deleted from $DEFAULT_PATH" 
-			underlines=$(echo "$str" | tr -c '\010' '-')
-			echo "$str\n$underlines"
+				cd $DEFAULT_PATH
+				rm -f *
+				clear
+				cd $DIR
+				str="All files deleted from $DEFAULT_PATH" 
+				underlines=$(echo "$str" | tr -c '\010' '-')
+				echo "$str\n$underlines"
+			
 		else
 			cd $DEFAULT_PATH
+
 			if [ -e $delete ]; then
 				rm $delete
 				clear
@@ -29,8 +37,9 @@ function delnote() {
 				underlines=$(echo "$str" | tr -c '\010' '-')
 				echo "$str\n$underlines"
 			else
-				echo "No such file."
+				echo "Invalid argument or parameter given."
 				cd $DIR
 		fi
 	fi
+fi
 }
